backend/pyproject.toml
[tool.poetry]
name = "forklore-backend"
version = "0.0.1"
description = "ForkLore Backend - Interactive Web Novel Platform"
authors = ["HueyJeong"]
readme = "README.md"
packages = [{include = "apps"}, {include = "config"}, {include = "common"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15"
djangorestframework-camel-case = "^1.4"
djangorestframework-simplejwt = "^5.3"
dj-rest-auth = {extras = ["with-social"], version = "^7.0"}
django-allauth = "^65.0"
drf-spectacular = "^0.27"
django-environ = "^0.11"
psycopg = {extras = ["binary"], version = "^3.2"}
pgvector = "^0.3"
celery = {extras = ["redis"], version = "^5.4"}
django-celery-beat = "^2.7"
gunicorn = "^23.0"
markdown = "^3.7"
google-generativeai = "^0.8"
Pillow = "^11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-django = "^4.9"
pytest-cov = "^6.0"
model-bakery = "^1.20"
factory-boy = "^3.3"
ipython = "^8.30"
django-debug-toolbar = "^4.4"
ruff = "^0.8"
pre-commit = "^4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config", "common"]
